feat: 알람 프롭스드롤링 및  리덕스 차이
