+++ src/components/header/HeaderMobile.jsx
@@ -1,0 +1,130 @@
+import React, { useState } from "react";
+import styled from "@emotion/styled";
+import alarmIcon from "../../assets/headerIcon/alarmIcon.svg";
+import profileIcon from "../../assets/headerIcon/profileIcon.svg";
+import logo from "../../assets/headerIcon/buycationLogo.webp";
+import logoHover from "../../assets/headerIcon/buycationLogoHover.webp";
+import ButtonBasic from "../elements/ButtonBasic";
+import Alarm from "./Alarm";
+import { getCookies } from "../../core/cookie";
+import { useNavigate } from "react-router-dom";
+import { useDispatch, useSelector } from "react-redux";
+import { sendModalStatus } from "../../redux/modules/modal/modalSlice";
+import { __patchAlarmState } from "../../redux/modules/alarm/alarmSlice";
+
+const HeaderMobile = (props) => {
+  const { onAlarmCount } = props;
+  const tokenValue = getCookies("id");
+  const navigate = useNavigate();
+  const dispatch = useDispatch();
+  const modalStatus = useSelector((state) => state.generalModal.toggleModal);
+  const [alarmModal, setAlarmModal] = useState(false);
+
+  //alarm
+  const onClickAlarmModalHandler = () => {
+    setAlarmModal(true);
+  };
+
+  const onMoveSelectPageHandler = (postingId, alarmId) => {
+    setAlarmModal(false);
+    navigate(`/details/${postingId}`);
+    dispatch(__patchAlarmState(alarmId));
+  };
+
+  const onCloseAlarmModalHandler = () => {
+    setAlarmModal(false);
+  };
+
+  //헤더 아이콘 기능
+  const onMoveLoginHandler = () => {
+    navigate("/login");
+    dispatch(sendModalStatus(true));
+  };
+
+  const onClickMypageModalHandler = () => {
+    dispatch(sendModalStatus(!modalStatus));
+  };
+
+  return (
+    <HeaderDiv>
+      {tokenValue ? (
+        <>
+          {alarmModal ? (
+            <Icon>
+              <img
+                className={onAlarmCount ? "mainColor" : ""}
+                alt="alarm"
+                src={alarmIcon}
+                onClick={onClickAlarmModalHandler}
+              />
+              <Alarm
+                top="4rem"
+                left="0"
+                onMove={onMoveSelectPageHandler}
+                onClose={onCloseAlarmModalHandler}
+              />
+            </Icon>
+          ) : (
+            <Icon>
+              <img
+                className={onAlarmCount ? "mainColor" : ""}
+                alt="alarm"
+                src={alarmIcon}
+                onClick={onClickAlarmModalHandler}
+              />
+            </Icon>
+          )}
+          <Logo alt="바이케이션" onClick={() => navigate("/")} />
+          <img
+            alt="tapBar"
+            src={profileIcon}
+            onClick={onClickMypageModalHandler}
+          />
+        </>
+      ) : (
+        <>
+          <Logo alt="바이케이션" onClick={() => navigate("/")} />
+          <ButtonBasic
+            width="4rem"
+            height="2rem"
+            borderRadius="2rem"
+            _onClick={onMoveLoginHandler}
+          >
+            로그인
+          </ButtonBasic>
+        </>
+      )}
+    </HeaderDiv>
+  );
+};
+
+export default HeaderMobile;
+
+const HeaderDiv = styled.div`
+  height: 60px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 2rem 1rem;
+  border-bottom: 1px solid ${({ theme }) => theme.colors.main};
+`;
+
+const Logo = styled.div`
+  height: 2.8rem;
+  width: 12.5rem;
+  margin-bottom: 1rem;
+  background: url(${logo}) bottom/100% no-repeat;
+  :hover {
+    background: url(${logoHover}) bottom/100% no-repeat;
+  }
+`;
+
+const Icon = styled.div`
+  display: flex;
+  gap: 23px;
+  cursor: pointer;
+
+  .mainColor {
+    filter: ${({ theme }) => theme.colors.imgFilter};
+  }
+`;
