+++ src/shared/layout/Header.jsx
@@ -1,19 +1,12 @@
 import React, { useEffect, useState } from "react";
-import styled from "@emotion/styled";
-import postingIcon from "../../assets/headerIcon/postingIcon.svg";
-import chattingIcon from "../../assets/headerIcon/chattingIcon.svg";
-import alarmIcon from "../../assets/headerIcon/alarmIcon.svg";
-import profileIcon from "../../assets/headerIcon/profileIcon.svg";
-import logo from "../../assets/headerIcon/buycationLogo.webp";
-import logoHover from "../../assets/headerIcon/buycationLogoHover.webp";
 import useWindowResize from "../../hooks/useWindowResize";
-import ButtonBasic from "../../components/elements/ButtonBasic";
+import UserModal from "../../components/header/UserModal";
+import HeaderPc from "../../components/header/HeaderPc";
+import HeaderMobile from "../../components/header/HeaderMobile";
 import { getCookies, removeCookies } from "../../core/cookie";
 import { useNavigate } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
-import UserModal from "../../components/header/UserModal";
 import { sendModalStatus } from "../../redux/modules/modal/modalSlice";
-import Alarm from "../../components/header/Alarm";
 import {
   __getAlarmCount,
   __patchAlarmState,
@@ -24,9 +17,9 @@ const Header = () => {
   const tokenValue = getCookies("id");
   const navigate = useNavigate();
   const dispatch = useDispatch();
+  const modalStatus = useSelector((state) => state.generalModal.toggleModal);
   const { alarmCount } = useSelector((data) => data?.alarm);
   const ALARMCOUNT = Number(alarmCount) >= 1;
-  const modalStatus = useSelector((state) => state.generalModal.toggleModal);
   const [alarmModal, setAlarmModal] = useState(false);
 
   //알람 갯수
@@ -36,7 +29,7 @@ const Header = () => {
     }
   }, [dispatch, tokenValue]);
 
-  //alarm
+  //alarm 조회, 삭제, 수정
   const onClickAlarmModalHandler = () => {
     setAlarmModal(true);
   };
@@ -49,19 +42,17 @@ const Header = () => {
     setAlarmModal(false);
   };
 
-  //헤더 아이콘 기능
+  //헤더 사람 아이콘 모달 내용 기능
   const onMovePostingHandler = () => {
     navigate("/posting");
     dispatch(sendModalStatus(true));
   };
+
   const onMoveMyProfileHandler = () => {
     navigate("/myprofile");
     dispatch(sendModalStatus(true));
   };
-  const onMoveLoginHandler = () => {
-    navigate("/login");
-    dispatch(sendModalStatus(true));
-  };
+
   const onMoveLogoutHandler = () => {
     removeCookies("id", {
       path: "/",
@@ -69,114 +60,27 @@ const Header = () => {
     navigate("/");
     dispatch(sendModalStatus(true));
   };
-  const onClickMypageModalHandler = () => {
-    dispatch(sendModalStatus(!modalStatus));
-  };
 
   return (
     <>
       {innerWidth > 768 ? (
-        <HeaderDiv>
-          <Logo alt="바이케이션" onClick={() => navigate("/")} />
-          {tokenValue ? (
-            <Icon>
-              <img
-                alt="posting"
-                src={postingIcon}
-                onClick={onMovePostingHandler}
-              />
-              <img alt="chatting" src={chattingIcon} />
-              {alarmModal ? (
-                <>
-                  <img
-                    className={ALARMCOUNT ? "mainColor" : ""}
-                    alt="alarm"
-                    src={alarmIcon}
-                    onClick={onClickAlarmModalHandler}
-                  />
-                  <Alarm
-                    top="4rem"
-                    right="0"
-                    onMove={onMoveSelectPageHandler}
-                    onClose={onCloseAlarmModalHandler}
-                  />
-                </>
-              ) : (
-                <img
-                  className={ALARMCOUNT ? "mainColor" : ""}
-                  alt="alarm"
-                  src={alarmIcon}
-                  onClick={onClickAlarmModalHandler}
-                />
-              )}
-              <img
-                alt="profile"
-                src={profileIcon}
-                onClick={onClickMypageModalHandler}
-              />
-            </Icon>
-          ) : (
-            <ButtonBasic
-              width="4rem"
-              height="2rem"
-              borderRadius="2rem"
-              _onClick={onMoveLoginHandler}
-            >
-              로그인
-            </ButtonBasic>
-          )}
-        </HeaderDiv>
+        <HeaderPc
+          onAlarmCount={ALARMCOUNT}
+          onClickAlarmModalHandler={onClickAlarmModalHandler}
+          onMoveSelectPageHandler={onMoveSelectPageHandler}
+          onCloseAlarmModalHandler={onCloseAlarmModalHandler}
+          onAlarmModal={alarmModal}
+        />
       ) : (
-        <HeaderDiv>
-          {tokenValue ? (
-            <>
-              {alarmModal ? (
-                <Icon>
-                  <img
-                    className={ALARMCOUNT ? "mainColor" : ""}
-                    alt="alarm"
-                    src={alarmIcon}
-                    onClick={onClickAlarmModalHandler}
-                  />
-                  <Alarm
-                    top="4rem"
-                    left="0"
-                    onMove={onMoveSelectPageHandler}
-                    onClose={onCloseAlarmModalHandler}
-                  />
-                </Icon>
-              ) : (
-                <Icon>
-                  <img
-                    className={ALARMCOUNT ? "mainColor" : ""}
-                    alt="alarm"
-                    src={alarmIcon}
-                    onClick={onClickAlarmModalHandler}
-                  />
-                </Icon>
-              )}
-              <Logo alt="바이케이션" onClick={() => navigate("/")} />
-              <img
-                alt="tapBar"
-                src={profileIcon}
-                onClick={onClickMypageModalHandler}
-              />
-            </>
-          ) : (
-            <>
-              <Logo alt="바이케이션" onClick={() => navigate("/")} />
-              <ButtonBasic
-                width="4rem"
-                height="2rem"
-                borderRadius="2rem"
-                _onClick={onMoveLoginHandler}
-              >
-                로그인
-              </ButtonBasic>
-            </>
-          )}
-        </HeaderDiv>
+        <HeaderMobile
+          onAlarmCount={ALARMCOUNT}
+          onClickAlarmModalHandler={onClickAlarmModalHandler}
+          onMoveSelectPageHandler={onMoveSelectPageHandler}
+          onCloseAlarmModalHandler={onCloseAlarmModalHandler}
+          onAlarmModal={alarmModal}
+        />
       )}
+
       {!modalStatus && (
         <UserModal
           top="4rem"
@@ -194,35 +98,3 @@ const Header = () => {
 };
 
 export default Header;
-
-const HeaderDiv = styled.div`
-  height: 60px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 2rem;
-  border-bottom: 1px solid ${({ theme }) => theme.colors.main};
-  @media screen and (max-width: 768px) {
-    padding: 2rem 1rem;
-  }
-`;
-
-const Logo = styled.div`
-  height: 2.8rem;
-  width: 12.5rem;
-  margin-bottom: 1rem;
-  background: url(${logo}) bottom/100% no-repeat;
-  :hover {
-    background: url(${logoHover}) bottom/100% no-repeat;
-  }
-`;
-
-const Icon = styled.div`
-  display: flex;
-  gap: 23px;
-  cursor: pointer;
-
-  .mainColor {
-    filter: ${({ theme }) => theme.colors.imgFilter};
-  }
-`;
